requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-environ==0.4.5
djangorestframework==3.12.4
googlemaps==4.4.5
haversine==2.3.0
idna==2.10
lxml==4.6.3
numpy==1.20.2
polyline==1.4.0
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
